Create the SpatialRuleIndex from Import and Web
