+++ config-example.properties
@@ -118,7 +118,7 @@ graph.dataaccess=RAM_STORE
 
 # Spatial Rules require you to provide Polygons in which the rules are enforced
 # The line below contains the default location for these rules
-# spatial_rules.location=web/src/main/resources/com/graphhopper/spatialrules/countries.geo.json
+# spatial_rules.location=core/files/spatialrules/countries.geo.json
 
 # You can define the maximum BBox for which spatial rules are loaded.
 # You might want to do this if you are only importing a small area and don't need rules for other countries.
