+++ core/files/spatialrules/README.MD
