+++ core/files/spatialrules/countries.geo.json
