+++ web/src/main/java/com/graphhopper/http/GraphHopperModule.java
@@ -20,8 +20,7 @@ package com.graphhopper.http;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.graphhopper.GraphHopper;
-import com.graphhopper.spatialrules.SpatialRuleLookupBuilder;
-import com.graphhopper.spatialrules.CountriesSpatialRuleFactory;
+import com.graphhopper.routing.util.spatialrules.SpatialRuleLookupBuilder;
 import com.graphhopper.GraphHopperAPI;
 import com.graphhopper.json.GHJson;
 import com.graphhopper.json.GHJsonFactory;
@@ -29,21 +28,17 @@ import com.graphhopper.json.geo.JsonFeatureCollection;
 import com.graphhopper.reader.osm.GraphHopperOSM;
 import com.graphhopper.routing.lm.LandmarkStorage;
 import com.graphhopper.routing.lm.PrepareLandmarks;
-import com.graphhopper.routing.util.DataFlagEncoder;
 import com.graphhopper.routing.util.EncodingManager;
-import com.graphhopper.routing.util.FlagEncoder;
-import com.graphhopper.routing.util.FlagEncoderFactory;
 import com.graphhopper.routing.util.spatialrules.DefaultSpatialRule;
 import com.graphhopper.routing.util.spatialrules.Polygon;
 import com.graphhopper.routing.util.spatialrules.SpatialRule;
 import com.graphhopper.routing.util.spatialrules.SpatialRuleLookup;
+import com.graphhopper.spatialrules.SpatialRuleLookupHelper;
 import com.graphhopper.storage.GraphHopperStorage;
 import com.graphhopper.storage.index.LocationIndex;
 import com.graphhopper.util.CmdArgs;
-import com.graphhopper.util.PMap;
 import com.graphhopper.util.Parameters;
 import com.graphhopper.util.TranslationMap;
-import com.graphhopper.util.shapes.BBox;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -110,28 +105,7 @@ public class GraphHopperModule extends AbstractModule {
             }
         }.forServer();
 
-        String spatialRuleLocation = args.get("spatial_rules.location", "");
-        if (!spatialRuleLocation.isEmpty()) {
-            try {
-                final BBox maxBounds = BBox.parseBBoxString(args.get("spatial_rules.max_bbox", "-180, 180, -90, 90"));
-                final FileReader reader = new FileReader(spatialRuleLocation);
-                final SpatialRuleLookup index = SpatialRuleLookupBuilder.buildIndex(new GHJsonFactory().create().fromJson(reader, JsonFeatureCollection.class), "ISO_A3", new CountriesSpatialRuleFactory(), .1, maxBounds);
-                logger.info("Set spatial rule lookup with " + index.size() + " rules");
-                final FlagEncoderFactory oldFEF = graphHopper.getFlagEncoderFactory();
-                graphHopper.setFlagEncoderFactory(new FlagEncoderFactory() {
-                    @Override
-                    public FlagEncoder createFlagEncoder(String name, PMap configuration) {
-                        if (name.equals(GENERIC)) {
-                            return new DataFlagEncoder(configuration).setSpatialRuleLookup(index);
-                        }
-
-                        return oldFEF.createFlagEncoder(name, configuration);
-                    }
-                });
-            } catch (IOException ex) {
-                throw new RuntimeException(ex);
-            }
-        }
+        SpatialRuleLookupHelper.buildAndInjectSpatialRuleIntoGH(graphHopper, args);
 
         graphHopper.init(args);
         return graphHopper;
